v0.2.1 - Improved CLI, and added support for package with different distribution than package names
For example, the yaml package for Python is called PyYAML (its package name), but its primary distribution name (the keyword used to import the package), is yaml. Those packages would not work, and the lambda packaging would try to install the non existing 'yaml' package. Now, the library will exchange the distribution name for the package name by using the importlib_metadata library.